changed num of coins to float in coinresponse

+++ PointGamingDesktop/POCO/CoinResponse.cs
@@ -8,7 +8,7 @@ namespace PointGaming.POCO
 	public class CoinResponse
 	{
 		public string _id { get; set; }
-		public int number_of_coins { get; set; }
+		public float number_of_coins { get; set; }
 		public string user_id { get; set; }
 		public string wallet_id { get; set; }
 	}
